sign-in & sign-up

+++ pkg/auth/token.go
@@ -8,11 +8,11 @@ import (
 )
 
 type TokenPayload struct {
-	UserID int `json:"user_id"`
+	UserID uint64 `json:"user_id"`
 }
 
 type Claims struct {
-	UserID int `json:"user_id"`
+	UserID uint64 `json:"user_id"`
 	jwt.RegisteredClaims
 }
